Fix an infinite recursion or a segfault in the lexer
The "external scanner" tries to detect eof by checking for a null byte, which has been ureliable in practice.
Sometimes the scanner loops forever, and sometimes it segfaults, when eof is reached without a null byte being reported.

The solution is to use the `eof()` predicate of the lexer which is more reliable.